chore(showcase): Add Roboto Studio site  to showcase (#21236)
Hi we're a Nottingham based dev studio, and we'd absolutely love it if you'd be able to showcase our site on the official Gatsby website. We're huge believers in Gatsby, to the degree that we have pivoted our entire dev process to focus on producing ultra-fast, SEO rich website using Gatsby.js and we're proactive in a lot of Headless CMS/JAMstack projects out there.

If there's any changes you'd love to see, drop a message across, we're always looking to improve the site.