documentation edit: Resolves #22580 (#22615)
* documentation edit: Resolves #22580

Changed comments in example code for the implementation of css in a custom cms module for the editor preview pane as per https://github.com/gatsbyjs/gatsby/issues/22580
Some of the information in the comment was erroneous.
So just bridging the gap in implementation/documentation for the editor preview pane in a custom cms module so people don't fall down a rabbit hole when deploying their site live on the netlify platform.

* Update README.md

Fixed typo

* chore: format

Co-authored-by: gatsbybot <mathews.kyle+gatsbybot@gmail.com>